% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-filedb.R
\name{deblob}
\alias{deblob}
\title{Convert serialized blob to R object}
\usage{
deblob(x, do_list = is.list(x))
}
\arguments{
\item{x}{serialized blob object.}

\item{do_list}{logical indicating whether to process \code{x} as a list.}

\item{connection, ascii}{arguments passed to \code{base::serialize()}.}
}
\value{
R object stored in a serialized blob.
}
\description{
Convert serialized blob to R object
}
