% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-filedb.R
\name{display_neurojam_schema}
\alias{display_neurojam_schema}
\title{Display neurojam database table schema}
\usage{
display_neurojam_schema(dbxcon, rankdir = "BT", apply_segments = FALSE, ...)
}
\arguments{
\item{dbxcon}{database connection}

\item{rankdir}{character value passed to
\code{datamodelr::dm_create_graph}.}

\item{apply_segments}{logical indicating whether tables should
be grouped into segments.}

\item{...}{additional arguments are ignored.}
}
\description{
Display neurojam database table schema
}
\details{
This function uses the \code{datamodelr} R package (if installed)
to produce a database table schema diagram showing the overall
database table structure, and intended field joins between tables.
}
