% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-filedb.R
\name{emblob}
\alias{emblob}
\title{Convert R object to serialized blob}
\usage{
emblob(x, connection = NULL, ascii = TRUE, do_list = FALSE)
}
\arguments{
\item{x}{R object}

\item{connection, ascii}{arguments passed to \code{base::serialize()}.}

\item{do_list}{logical indicating whether to process \code{x} as a list.}
}
\value{
serialized blob objects sufficient to store in a relational
database. If input \code{x} was a list, the output will be a list
of serialized blob objects.
}
\description{
Convert R object to serialized blob
}
