% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-baseline.R
\name{extract_baseline_data}
\alias{extract_baseline_data}
\title{Extract animal baseline signal data from raw data}
\usage{
extract_baseline_data(
  con,
  animal = NULL,
  project = NULL,
  phase = NULL,
  source_filename = NULL,
  source_dirname = NULL,
  rdata_filename = NULL,
  rdata_dirname = NULL,
  animal_raw_data_table = "animal_raw_data",
  before_event_only = TRUE,
  baseline_step = 60,
  baseline_max_duration = 480,
  all_start_grep = "^Start",
  all_start_row = 2,
  all_start_fixed = NULL,
  event_grep = "^EV",
  event_start = 2,
  attr_names = c("animal", "channel", "ts.step", "ts"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{before_event_only}{logical indicating whether to
require all baseline data to occur before the first
event time stamp.}

\item{baseline_step}{numeric value indicating the time bins
to return for the baseline. The total baseline signal
will be split into parts of this duration. The time
is assumed to have units of \code{"seconds"}, consistent
with units recorded in the time \code{"step"} field.}

\item{baseline_max_duration}{numeric value indicating
the maximum duration of signal to be obtained for
baseline signal. This argument is useful when the
baseline may be recorded for an extended length of
time, but only a certain duration is sufficient for
analysis. Time units are assumed to be \code{"seconds"},
consistent with the time \code{"step"} field.}

\item{all_start_grep}{character string used to identify
the element name containing a matrix of start times.
When supplied, \code{all_start_row} is used to determine
the row indicating the start of the baseline signal.}

\item{all_start_row}{integer value indicating the row
containing the experiment start time index.}

\item{all_start_fixed}{numeric value indicating the
start time index, used as an optional replacement
of the \code{all_start_grep}, and \code{all_start_row} arguments.}

\item{event_grep}{character pattern used to identify
element names that contain event time stamps.}

\item{event_start}{integer value indicating the element
among results matched by \code{event_grep} to use for the
event start times. For example, when event names
are \code{c("EVT01", "EVT02", "EVT03")} and \code{"EVT02"}
contains start times, use \code{event_start=2}.
This value is relevant only when \code{before_event_only=TRUE}.}

\item{attr_names}{character vector of attribute names which
are present in the raw data for each channel, which will
be retained in the resulting baseline signal data.}
}
\description{
Extract animal baseline signal data from raw data
}
\details{
This function uses \code{get_animal_raw_data()} to query the
database, retrieve the associated animal raw data, then
extracts baseline signal typically before the first
annotated event. This baseline signal can be useful
for calculations that measure the change in frequency
across events.
}
