% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-filedb.R
\name{extract_event_db}
\alias{extract_event_db}
\title{Extract animal event data from database}
\usage{
extract_event_db(
  con,
  filename = NULL,
  file_table = "ephys_file",
  file_event_table = "file_event",
  file_channel_table = "file_channel",
  file_channel_event_table = "file_channel_event",
  event_prestart = 4,
  event_poststop = 4,
  event_type = "event",
  default_event_diff = 40,
  load_db = TRUE,
  reuse_db = TRUE,
  verbose = FALSE,
  ...
)
}
\description{
Extract animal event data from database
}
