% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-db.R
\name{get_animal_event_derived}
\alias{get_animal_event_derived}
\title{Get animal event derived data}
\usage{
get_animal_event_derived(
  con,
  animal = NULL,
  project = NULL,
  phase = NULL,
  step = NULL,
  channel = NULL,
  event = NULL,
  event_prestart = NULL,
  event_poststop = NULL,
  derived_type = NULL,
  derived_extra = NULL,
  animal_event_derived_table = "animal_event_derived",
  return_data = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{con}{database connection with class \code{DBIConnection},
for example produced from \code{DBI::dbConnect()}. The database
can be of any type compatible with \code{DBI}, for example
\code{dbConnect(RMySQL::MySQL())} produces an object of
class \code{MySQLConnection} which extends \code{DBIConnection}.}

\item{animal, project, phase, step, channel, event, event_prestart, event_poststop}{arguments used to identify the unique entry from which
the event data was derived, equivalent to parameters
passed to \code{get_animal_event_data()}.}

\item{derived_type}{character string describing the type
of data being stored, used in subsequent queries.}

\item{derived_extra}{character string intended to describe
any additional arguments or parameters related to the
method used to derive results.}

\item{animal_event_derived_table}{character string with
the database table in which to store the derived result
data.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Get animal event derived data
}
\details{
This function returns data stored by \code{save_animal_event_derived()}
from a relational database. The derived data is returned
when argument \code{return_data=TRUE}, in the form of a \code{list} column
of a \code{data.frame}.
}
\seealso{
Other jam database functions: 
\code{\link{ephys_short_name}()},
\code{\link{extract_event_data}()},
\code{\link{get_animal_event_data}()},
\code{\link{get_animal_raw_data}()},
\code{\link{save_animal_event_data}()},
\code{\link{save_animal_event_derived}()},
\code{\link{save_animal_raw_data}()},
\code{\link{sqlite_desc}()},
\code{\link{sqlite_indices}()},
\code{\link{sqlite_view}()}
}
\concept{jam database functions}
