% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-db.R
\name{get_animal_raw_data}
\alias{get_animal_raw_data}
\title{Save raw signal data per animal}
\usage{
get_animal_raw_data(
  con,
  animal = NULL,
  project = NULL,
  phase = NULL,
  source_filename = NULL,
  source_dirname = NULL,
  rdata_filename = NULL,
  rdata_dirname = NULL,
  animal_raw_data_table = "animal_raw_data",
  channel_grep = "^AI[0-9]+$|^FP[0-9]+",
  return_rdata = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Save raw signal data per animal
}
\seealso{
Other jam database functions: 
\code{\link{ephys_short_name}()},
\code{\link{extract_event_data}()},
\code{\link{get_animal_event_data}()},
\code{\link{get_animal_event_derived}()},
\code{\link{save_animal_event_data}()},
\code{\link{save_animal_event_derived}()},
\code{\link{save_animal_raw_data}()},
\code{\link{sqlite_desc}()},
\code{\link{sqlite_indices}()},
\code{\link{sqlite_view}()}
}
\concept{jam database functions}
