% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-import.R
\name{import_ephys_mat}
\alias{import_ephys_mat}
\title{Import Plexon MatLab data}
\usage{
import_ephys_mat(
  mat_files_use,
  animal = NULL,
  phase = NULL,
  channels_per_animal = 5,
  channels_to_retain = seq_len(channels_per_animal - 1),
  drop_unused_channels = TRUE,
  do_plot = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{mat_files_use}{data.frame containing columns described
above, used to select the appropriate MatLab .mat file
to import.}

\item{animal}{character vector with one animal identifier}

\item{phase}{character vector with one phase}

\item{channels_per_animal}{integer number of channels to assign for
each animal.
If 5 channels were assigned per animal,
\code{channels_per_animal=5} which
will assign blocks of 5 channels to each animal identifier
associated with the .mat file.}

\item{channels_to_retain}{integer vector of channels to retain for
each animal. For example, there may be 5 channels assigned
to each animal, and \code{channels_to_retain} defines which of
those channels should be used for downstream analysis.
If 5 channels were assigned per animal, \code{channels_per_animal=5} which
will assign channels to each animal in blocks of 5. However,
if only 4 channels were enabled during recording,
\code{channels_to_retain=c(1,2,3,4)} will cause the first 4 channels to
be retained.}

\item{drop_unused_channels}{logical indicating whether to drop
all channels except those assigned to the animal, and
channels to be retained as defined by \code{channels_to_retain}.}

\item{do_plot}{logical indicating whether to call
\code{plot_ephys_event_data()} to plot an overview of the channel
signal, and the signal defined for each event.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{plot_ephys_event_data()}
when \code{do_plot=TRUE}.}
}
\description{
Import MatLab .mat files specifically exported from Plexon
NeuroExplorer software.
}
\details{
The \code{mat_files_use} data.frame should contain the following colnames:
\itemize{
\item path: character column containing the full file path to each .mat file
\item filename: character column containing the base filename,
without folder or path information.
\item animal: character column with animal identifier, where multiple
animals are separated by a space.
\item phase: character column containing phase, used to delineate multiple
experimental phases.
}
}
\seealso{
Other jam import functions: 
\code{\link{import_ephys_mat_1}()}
}
\concept{jam import functions}
