% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-matrix.R
\name{list2groups}
\alias{list2groups}
\title{Convert list to groups}
\usage{
list2groups(x, ...)
}
\description{
Convert list to groups
}
\details{
This function simply expands the \code{names(x)} to \code{lengths(x)}
which assigns group labels to each element in \code{x}.
}
