% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-biwavelet.R
\name{summarize_event_bins}
\alias{summarize_event_bins}
\title{Summarize event bins for a numeric matrix}
\usage{
summarize_event_bins(
  x,
  col_values = NULL,
  pre_time = 0,
  post_time = 0,
  pre_bin_n = 1,
  time_bin_n = 2,
  post_bin_n = 1,
  pre_bin_labels = "pre",
  time_bin_labels = c("tone", "trace"),
  post_bin_labels = "post",
  useMedian = FALSE,
  trim = NULL,
  make_tall = FALSE,
  adj_freq_range = c(3, 5),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric matrix whose \code{colnames(x)} represent time.}

\item{pre_time}{numeric value indicating the time before the
event occurs. By default, the event is assumed to occur
at time=0, so the lowest negative value is used to
derive \code{pre_time}.}

\item{post_time}{numeric value indicating the time added
to the end of the event, used as a post-event duration.}

\item{pre_bin_n, time_bin_n, post_bin_n}{integer number of bins
used to sub-divide each section, for pre-event, event,
and post-event, respectively.}

\item{pre_bin_labels}{character vector used to label the
pre-event bin sections, expanded to the number of pre-event bins.}

\item{time_bin_labels}{character vector used to label the
post-event bin sections, expanded to the number of post-event bins.}

\item{useMedian}{logical passed to \code{jamba::rowGroupMeans()}
indicating whether to use the row median value to summarize
each bin.}

\item{make_tall}{logical indicating whether to convert data
to tall \code{data.frame} format. When \code{make_tall=FALSE} the
data is returned as a numeric matrix.}

\item{adj_freq_range}{numeric range of frequencies to use
as a baseline, which is subtracted from the resulting
profile for each time bin. This subtraction is only
applied when \code{make_tall=TRUE}, and
\code{length(adj_freq_range) > 0}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\value{
numeric \code{matrix} whose colnames are the event bins
after being sub-divided, when \code{make_tall=FALSE}. When
\code{make_tall=TRUE} the data is returned as a \code{data.frame}
with each row containing one frequency in column
\code{"value"}.
}
\description{
Summarize event bins for a numeric matrix
}
\details{
This function is intended to summarize event data stored
in a numeric matrix with time columns, based upon
pre-event, event, and post-event sections. Each section
can be subdivided into \code{n} parts, using arguments
\code{pre_bin_n}, \code{time_bin_n}, and \code{post_bin_n}, respectively.

Note that the input numeric matrix \code{x} should already
be transformed appropriate for summarization. For example,
if \code{log2} transformation is required, it should be done
before calling this function.
}
