% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-biwavelet.R
\name{calc_ephys_wavelet}
\alias{calc_ephys_wavelet}
\title{Calculate wavelet transformation}
\usage{
calc_ephys_wavelet(x, iWt = NULL, return_type = c("wt", "m", "list"),
  type = c("power.corr.norm", "power", "power.corr", "power.norm",
  "wavelet", "phase"), x_condense_factor = 10, x_label_multiple = 5,
  i_step = 0.001, m = NULL, do.sig = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{numeric matrix with two columns, where the first column
contains neural signal data, and the second column contains
timestamps in units of seconds.
When \code{iWt} is supplied, \code{x} is not used.}

\item{iWt}{list output from \code{biwavelet::wt()}, or NULL.
When \code{iWt} is supplied, \code{x} is not used. When \code{iWt} is not
supplied, it is only calculated when \code{return_type} is
either \code{"wt"} or \code{"list"}, or when \code{"m"} is not supplied.}

\item{return_type}{character vector indicating the type of
data to return: \code{"wt"} returns the full output of \code{biwavelet::wt()};
\code{"m"} returns the numeric frequency matrix; \code{"list"} returns
a list including elements \code{c("wt", "m")} which include the
previous two formats.}

\item{type}{character indicating which frequency matrix to
extract from the output of \code{biwavelet::wt()}. Note that
when \code{iWt} is supplied, the matrix is taken directly from it.
When \code{m} is supplied as a numeric matrix, it is used
directly, and \code{type} is not used.}

\item{x_condense_factor}{integer value indicating the level
of compression to condense the numeric matrix, when \code{return_type}
is either \code{"m"} or \code{"list"}. When \code{x_condense_factor=1} there
is no compression, otherwise it calls \code{condense_freq_matrix()}.}

\item{x_label_multiple}{integer value indicating the multiple
to use when defining labels to display. For example
when \code{x_label_multiple=5} labels are displayed at multiples
of 5: 0, 5, 10, 15, 20, 25.}

\item{m}{numeric matrix, optionally supplied as an alternative
to extracting a numeric matrix from either \code{iWt} or
the output of \code{biwavelet::wt(x)}.}

\item{do.sig}{logical indicating whether to calculate
statistical significance. This argument is passed to
\code{biwavelet::wt()}.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{biwavelet::wt()}
when \code{iWt} is not supplied.}
}
\value{
When \code{return_type} is \code{"wt"} the output of
\code{biwavelet::wt()} is returned; when \code{return_type} is
\code{"m"} a numeric matrix is returned, whose rownames
are the \code{"period"} and colnames are the timestamp,
this matrix may also be condensed when \code{x_condense_factor}
is greater than 1; when \code{return_type} is \code{"list"} then
a list is returned with elements \code{"wt"} and \code{"m"}
containing the two previous return types. The numeric
matrix \code{"m"} has usseful attributes: \code{"lab_col"} is
a character vector of column labels at user-friendly
positions; \code{"lab_row"} is a character vector of row
labels at user-friendly positions, after converting
\code{"period"} to frequency in Hertz (per second) units.

Since the data matrix is very large, and it is often
not necessary to retain the full resolution of this data,
it can be condensed by setting \code{x_condense_factor} to any
value above 1. In that case \code{condense_freq_matrix()} is
called to condense the numeric matrix to a smaller size
before returning results. This process reduces data size
substantially, while avoiding memory consumption of returning
each numeric matrix produced by \code{biwavelet::wt()}.
}
\description{
This function takes numeric signal data, and performs
biwavelet transformation by calling \code{biwavelet::wt()}.
It optionally returns a specific numeric matrix,
for example when \code{type="power.corr.norm"} it returns
the normalized power correlation matrix (see docs for
\code{biwavelet::wt()} for more details).
}
\details{
This function attempts to return a numeric spectral
frequency matrix, either by calling \code{biwavelet::wt()}
then extracting the appropriate numeric matrix from
the resulting list object; or by using the \code{iWt} data
from a previous call to \code{biwavelet::wt()} to
extract the appropriate numeric matrix; or by using
a numeric matrix \code{m} supplied directly. In the latter
case, \code{biwavelet::wt()} is not necessary and may
be substantially faster.
}
