% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-biwavelet.R
\name{choose_exemplar_channel}
\alias{choose_exemplar_channel}
\title{Choose exemplar channel}
\usage{
choose_exemplar_channel(x, exemplar_method = c("highest", "median",
  "mean"), group_colnames = c("time_bin"), animal_colname = "animal",
  event_colname = "event", channel_colname = "channel",
  value_colname = "value", freq_colname = "freq", freq_range = c(7,
  15), ...)
}
\description{
Choose exemplar channel
}
\details{
This function takes tall output from \code{summarize_event_bins()},
with multiple channels and events per animal, and choose one
channel to serve as the exemplar for analysis for each animal,
across all events.

The basic assumption is that each channel is representing the
same signal, but with different level of specificity or
signal-to-noise, dependent upon the sensor associated with
each channel. This function attempts to choose the one channel
with highest signal, and best signal-to-noise.

Alternatively, this function can be used to calculate the
mean or median signal across all channels.
}
