% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-matrix.R
\name{condense_freq_matrix}
\alias{condense_freq_matrix}
\title{Condense a spectral frequence data matrix}
\usage{
condense_freq_matrix(x, column_n = ncol(x), row_n = nrow(x),
  verbose = TRUE, ...)
}
\description{
Condense a spectral frequence data matrix
}
