% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-matrix.R
\name{cutIntoChunks}
\alias{cutIntoChunks}
\title{Cut vector into a list of fixed size}
\usage{
cutIntoChunks(x, n = 1, ...)
}
\arguments{
\item{x}{\code{vector}, or any R object compatible with \code{base::split()}.}

\item{n}{integer number of list elements to create.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{list} of vectors with length \code{n}.
}
\description{
Cut vector into a list of fixed size
}
\details{
This function provides a basic method to split a vector
into a list of vectors of a known length, with each list element
being roughly the same size.
}
\examples{
x <- 1:170;
cutIntoChunks(x, n=4);
lengths(cutIntoChunks(x, n=8))

}
\seealso{
Other jam matrix functions: \code{\link{condense_freq_matrix}},
  \code{\link{discretize_labels}},
  \code{\link{freq_heatmap}},
  \code{\link{signal_freq_heatmap}}
}
\concept{jam matrix functions}
