% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-matrix.R
\name{discretize_labels}
\alias{discretize_labels}
\title{Discretize labels}
\usage{
discretize_labels(x, target = NULL, pretty.n = 15, ...)
}
\description{
Discretize labels
}
\details{
This function takes a vector of numeric values, a vector of
target values intended to be used as labels, and returns
a vector of labels where all unmatched target values are
replaced with \code{""} blank, and matched target values use
the target label.
}
\examples{
x <- seq(from=-2, to=2, by=0.2);
x;
xd <- discretize_labels(x, target=pretty(x));
data.frame(original_value=x, discretized_label=xd);

}
\seealso{
Other jam matrix functions: \code{\link{condense_freq_matrix}},
  \code{\link{cutIntoChunks}}, \code{\link{freq_heatmap}},
  \code{\link{signal_freq_heatmap}}

Other jam heatmap functions: \code{\link{freq_heatmap}},
  \code{\link{signal_freq_heatmap}}
}
\concept{jam heatmap functions}
\concept{jam matrix functions}
