% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-biwavelet.R
\name{event_freq_profile}
\alias{event_freq_profile}
\title{Event to frequency profile}
\usage{
event_freq_profile(events_m_1, i_step = 0.001, verbose = FALSE, ...)
}
\description{
Event to frequency profile
}
