% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-db.R
\name{extract_event_data}
\alias{extract_event_data}
\title{Extract animal event data from raw data}
\usage{
extract_event_data(con, animal = NULL, project = NULL, phase = NULL,
  source_filename = NULL, source_dirname = NULL,
  rdata_filename = NULL, rdata_dirname = NULL,
  animal_raw_data_table = "animal_raw_data", event_prestart = 4000,
  event_poststop = 4000, verbose = FALSE, ...)
}
\description{
Extract animal event data from raw data
}
\details{
This function uses \code{get_animal_raw_data()} to
query the database, retrieve the associated
animal raw data, then extracts the available
event data.
}
