% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-classes.R
\name{FreqTimeMatrix}
\alias{FreqTimeMatrix}
\title{Create FreqTimeMatrix numeric matrix with frequency rows and time columns}
\usage{
FreqTimeMatrix(x, row_units = "Hz", row_title = "frequency",
  row_starts = as.numeric(rownames(x)), row_ends = NULL,
  row_step = NULL, column_title = "time", column_units = "sec",
  column_starts = as.numeric(colnames(x)), column_ends = NULL,
  column_step = NULL, verbose = FALSE, ...)
}
\description{
Create FreqTimeMatrix numeric matrix with frequency rows and time columns
}
\details{
This function creates an object class \code{FreqTimeMatrix} which extends
\code{data.table} and represents frequency in rows, and time in columns.
It was created to maintain proper numeric values associated with
rows and columns, and specifically the range of values represented.

For example, consider columns whose values are \code{c(0.000, 0.001, 0.002)}.
The first column represents a time range from \code{0.000} until \code{0.001},
the second column represents a time range from \code{0.001} until \code{0.002},
and so on. The duration represented by each column is inferred only
by the presence of additional column with consistent step differences.
If we extracted only the first column of data, there would be no
way of knowing the range of time represented.

Further, if we want to condense all points into 1-second intervals,
we would want all time ranges from \code{0.000} until \code{1.000}, but not
including the time range from \code{1.000} until \code{1.001}. Extracting exactly
the appropriate columns is straightforward. But the output matrix will
have columns that represent 1000-fold longer time range, so the output
matrix will also need to represent the new time range.

To summarize:
\itemize{
\item FreqTimeMatrix represents row and column numeric values as
\code{row_starts}, \code{column_starts}, which indicate the starting values;
similarly \code{row_ends}, \code{column_ends} indicate the ending values.
\item FreqTimeMatrix represents row and column units, simply as a text
label. No further unit-based operations are supported yet.
\item FreqTimeMatrix data can be summarized by groups of columns or
rows, and the output FreqTimeMatrix should indicate the new
range of numeric values represented, as relevant.
}
}
