% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-db.R
\name{get_animal_event_data}
\alias{get_animal_event_data}
\title{Get animal event data}
\usage{
get_animal_event_data(con, animal = NULL, project = NULL,
  phase = NULL, channel = NULL, event = NULL,
  return_signal = FALSE, ...)
}
\description{
Get animal event data
}
