% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-db.R
\name{get_animal_raw_data}
\alias{get_animal_raw_data}
\title{Save raw signal data per animal}
\usage{
get_animal_raw_data(con, animal = NULL, project = NULL, phase = NULL,
  source_filename = NULL, source_dirname = NULL,
  rdata_filename = NULL, rdata_dirname = NULL,
  animal_raw_data_table = "animal_raw_data", return_rdata = FALSE,
  verbose = FALSE, ...)
}
\description{
Save raw signal data per animal
}
