% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-events.R
\name{get_ephys_event_data}
\alias{get_ephys_event_data}
\title{Get Ephys event data}
\usage{
get_ephys_event_data(mat_l, channel, events = NULL,
  event_offset = NULL, event_start = 2, event_stop = 3,
  event_prestart = 4000, event_poststop = 4000, event_label = "Tone",
  event_grep = "^EV", attr_names = c("animal", "channel", "ts.step",
  "ts"), verbose = FALSE, ...)
}
\description{
Get Ephys event data
}
\seealso{
Other jam event functions: \code{\link{plot_ephys_event_data}}
}
\concept{jam event functions}
