% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-filedb.R
\name{import_ephys_file}
\alias{import_ephys_file}
\title{Import Ephys raw data file into relational database}
\usage{
import_ephys_file(filename, con, mat_l = NULL, project = "",
  phase = "", file_table = "ephys_file",
  file_event_table = "file_event", file_channel_table = "file_channel",
  channel_grep = "^(AI|FP)(0[1-4]|1[7-9]|2[0])$",
  time_step_grep = "[.]ts[.]step$", index_offset_grep = "[.]ind$",
  event_grep = "^EVT[0-9]+$", drop_grep = "^(WB|SP|AI|FP).*$",
  start_grep = "^Start$", stop_grep = "^Stop$",
  duration_grep = "^AllFile$", valid = TRUE, override = FALSE,
  override_method = c("invalidate", "delete"), verbose = TRUE, ...)
}
\arguments{
\item{filename}{character string or vector of file names to import.
Duplicate \code{filename} values are not permitted. If a duplicate
\code{filename} is intended to be loaded as a separate entity,
include all or path of the file path in order to make the
\code{filename} character string unique in the database.}

\item{con}{database connection as produced by \code{DBI::dbConnect()}
or compatible database-specific package \code{RSQLite::dbConnect()}.}

\item{mat_l}{optional R object containing data imported from
\code{filename}, used to save time re-importing the same file if
the data had already previously been loaded. Note that
\code{filename} is always required, since the \code{filename} is
the primary key for stored data. As such, duplicate \code{filename}
values are not permitted. Note that \code{mat_l} may also be
an environment containing the relevant data.}

\item{file_table}{character name of the database table to store
Ephys file data.}

\item{file_event_table}{character name of the database table to store
the file-to-event relationship. Events are stored for each
file, and are separately applied to each channel in the file.}

\item{file_channel_table}{character name of the database table to store
the signal for each channel that meets criteria by \code{channel_grep}.}

\item{channel_grep}{regular expression pattern used to identify
measurement channels in the file. Any event names matched
by \code{event_grep} are removed from the channel list.}

\item{time_step_grep}{regular expression pattern used to recognize
the name of the time step, expected to be a suffix of
the channel name. For example channel \code{"AI01"} would imply
a time step name \code{"AI01.ts.step"}, from which the time step
is obtained.}

\item{index_offset_grep}{regular expression pattern used to recognize
the name of the index offset, expected to be a suffix of
the channel name. For example, channel \code{"AI01"} would imply
an index offset name \code{"AI01.ind"}.}

\item{event_grep}{regular expression pattern used to identify
names of events in the file. Note that event names take
priority over channel names.}

\item{drop_grep}{regular expression pattern used to exclude
certain channel names from being stored.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are ignored.}
}
\description{
Import Ephys raw data file into relational database
}
