% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-import.R
\name{import_ephys_mat_1}
\alias{import_ephys_mat_1}
\title{Import single MatLab .mat file}
\usage{
import_ephys_mat_1(mat_file, animal, channels_per_animal = 5,
  channels_to_retain = seq_len(channels_per_animal - 1),
  drop_unused_channels = TRUE, verbose = FALSE, ...)
}
\description{
Import single MatLab .mat file
}
\seealso{
Other jam import functions: \code{\link{import_ephys_mat}}
}
\concept{jam import functions}
