% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-misc.R
\name{list2igraph}
\alias{list2igraph}
\title{Convert list to igraph}
\usage{
list2igraph(x, delim = "[, \\t]+", type = c("star", "full"),
  color_sub = c(names = "dodgerblue", values = "gold"), useGrey = 15,
  ...)
}
\arguments{
\item{x}{list object, or character vector which will be split
into a list using \code{base::strsplit()} in the form \code{strsplit(x, delim)}.}

\item{delim}{character value used when \code{x} is not a list, passed
to \code{base::strsplit()}.}

\item{type}{character string indicating the type of network defined
for each list element: \code{"star"} creates a star network, each
node is connected to a central node; \code{"full"} creates a fully
connected network with each node connected to all other nodes.}

\item{color_sub}{named vector of colors, either containing names
\code{"names"} and \code{"values"} for default colors for the \code{names(x)}
and the values in \code{x}, respectively; or a named vector whose
names are \code{names(x)} and/or values in \code{x}.}

\item{...}{additional arguments are passed to
\code{jamba::setTextContrastColor()} to determine a contrasting
label color.}
}
\description{
This function takes a list of character vectors, and converts
them to an igraph object. It is intended to help recognize when
multiple list elements share the same items.
}
\details{
When \code{type} is \code{"full"} each list element is converted to
a fully connected network (all entries connected to all other
entries.)

When \code{type} is \code{"star"} each list element is converted to
a star network, with each element connected to one central
node names(x).
}
\examples{
x <- list(
   A=c("one", "two", "three"),
   B=c("one", "four", "five"),
   C=c("six", "seven", "eight")
);
plot(list2igraph(x))

}
