% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-biwavelet.R
\name{matrix_period2hz}
\alias{matrix_period2hz}
\title{Convert matrix scale from period to Hertz units}
\usage{
matrix_period2hz(x, row_values = NULL, freq_range = c(1, 20),
  freq_step = 0.2, freq_method = c("2"), do_rev = FALSE,
  verbose = FALSE, ...)
}
\value{
numeric matrix whose rows are in units of frequency
in Hertz, observations per second. An attribute \code{"row_values"}
contains a numeric vector for each row.
}
\description{
Convert matrix scale from period to Hertz units
}
\details{
This function converts a numeric matrix whose rows are
in units of period, number of seconds per observation,
into units of Hertz, number of observations per second.
It also makes the Hertz scale linear by using linear
interpolation between actual period measurement values,
via the \code{stats::approx()} function.

The period values are expected to be \code{rownames(x)},
which are converted to numeric with \code{as.numeric(rownames(x))}.
Otherwise the argument \code{row_values} can be used to supply
a numeric vector.
}
