% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-biwavelet.R
\name{plot_ephys_wavelet}
\alias{plot_ephys_wavelet}
\title{Plot Ephys wavelet data}
\usage{
plot_ephys_wavelet(x, iWt = NULL, style = c("default", "jam", "CH"),
  type = c("power.corr.norm", "power", "power.corr", "power.norm",
  "wavelet", "phase"), plot.coi = FALSE, plot.cb = FALSE,
  plot.phase = FALSE, plot.sig = FALSE, col.coi = "grey30",
  fill.cols = warpRamp(getColorRamp("RdBu_r"), lens = 2),
  ylim = rev(2^c(0, -6)), arrow.col = "green4", col = "RdBu_r",
  freq_lim = NULL, i_step = 0.001, x_compress_factor = 10,
  do_plot = TRUE, m = NULL, ...)
}
\description{
Plot Ephys wavelet data
}
