% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-db.R
\name{save_animal_event_data}
\alias{save_animal_event_data}
\title{Save event signal data per animal}
\usage{
save_animal_event_data(con, events_m, animal, project, phase = "",
  step = NULL, animal_event_data_table = "animal_event_data", ...)
}
\arguments{
\item{con}{database connection with class \code{DBIConnection},
for example produced from \code{DBI::dbConnect()}. The database
can be of any type compatible with \code{DBI}, for example
\code{dbConnect(RMySQL::MySQL())} produces an object of
class \code{MySQLConnection} which extends \code{DBIConnection}.}

\item{events_m}{matrix containing lists of numeric vectors,
where each cell contains a numeric vector of raw signal.
The \code{colnames(events_m)} are channel names, and
\code{rownames(events_m)} are event names.}
}
\description{
Save event signal data per animal
}
\details{
This function saves the subset of raw signal per
per animal, per channel, and per annotated event.
}
