% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-matrix.R
\name{signal_freq_heatmap}
\alias{signal_freq_heatmap}
\title{Heatmap of frequency-time matrix for animal signal data}
\usage{
signal_freq_heatmap(dbxcon, animal = NULL, project = NULL,
  phase = NULL, filename = NULL, channel = NULL,
  time_step_sec = NULL, freq_step_hz = NULL, plot = FALSE,
  type = c("df", "hm", "Heatmap", "matrix", "m"),
  include_motion = FALSE, motion_table = "animal_motion_data",
  motion_column = "freeze_cnt", top_annotation = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{dbxcon}{DBI database connection}

\item{animal, project, phase, filename, channel, time_step_sec, freq_step_hz}{arguments used to filter the database table \code{fca_freq_matrix} in
order to specify only one row in the results, therefore only one
frequency-time matrix.}

\item{plot}{logical indicating whether to plot the output. When
\code{plot=FALSE} the \code{data.frame} of query results is returned, and
no heatmap is created. This option can be helpful when querying
the database.}

\item{type}{character value indicating the type of output, where
\code{"Heatmap"} will return the Heatmap object; and \code{"matrix"} will
return the numeric matrix itself, and not the heatmap.}

\item{include_motion}{logical indicating whether to query the database
for table \code{"animal_motion_data"} to retrieve motion data for the
given \code{animal} and \code{phase}. When the database table exists,
\code{top_annotation} will be created using
\code{ComplexHeatmap::anno_barplot()} to produce a barplot at the
top of the heatmap. For more customization, use \code{top_annotation}.}

\item{motion_table, motion_value_column}{optional character strings with
the database table and table column to use to retrieve motion
data. Note that the table is expected to contain columns:
\code{"animal","phase","time_sec"} as well as the \code{motion_value_column}.
The \code{"time_sec"} values are expected to match the time values
in the \code{"fca_freq_matrix"} table.}

\item{top_annotation}{optional \code{HeatmapAnnotation} object as produced
by \code{ComplexHeatmap::HeatmapAnnotation()}, suitably aligned with
the columns of data included in the heatmap. This annotation
will be subsetted by \code{freq_heatmap()} as relevant when
\code{column_range} is passed to that function.}

\item{verbose}{logical indicating whether to print verbose output.}

\item{...}{additional arguments are passed to \code{freq_heatmap()},
see that function documentation for info on customizing the
heatmap.}
}
\description{
Heatmap of frequency-time matrix for animal signal data
}
\details{
This function queries the database to return data, then
calls \code{freq_heatmap()} to create the heatmap.
}
\examples{
# db_path <- "ephys_db.sqlite";
# dbxcon <- dbx::dbxConnect(adapter="sqlite", dbname=db_path);

## Basic query of available data
# head(signal_freq_heatmap(dbxcon), 20)

## Tabulate data by phase
# table(signal_freq_heatmap(dbxcon)[,c("project","phase")])
# table(signal_freq_heatmap(dbxcon)[,c("animal","phase")])

## Provide animal and phase, create heatmap
# signal_freq_heatmap(dbxcon, animal="AF11-4", phase="Acquisition", plot=TRUE)

## Another example
# signal_freq_heatmap(dbxcon, animal="SA88-1", phase="Acquisition", plot=TRUE)

## Zoom into a specific time range
# signal_freq_heatmap(dbxcon, animal="SA88-1", phase="Acquisition", plot=TRUE, column_range=c(400,700))

}
\seealso{
Other jam matrix functions: \code{\link{condense_freq_matrix}},
  \code{\link{cutIntoChunks}},
  \code{\link{discretize_labels}},
  \code{\link{freq_heatmap}}

Other jam heatmap functions: \code{\link{discretize_labels}},
  \code{\link{freq_heatmap}}
}
\concept{jam heatmap functions}
\concept{jam matrix functions}
