% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurojam-sqlite.R
\name{sqlite_indices}
\alias{sqlite_indices}
\title{Get SQLite table index information}
\usage{
sqlite_indices(con, tbl_name = NULL, ...)
}
\arguments{
\item{con}{valid SQLite database connection}

\item{tbl_name}{character vector of tables to return, or when
\code{NULL} all tables are returned.}

\item{...}{additional arguments are ignored.}
}
\value{
\code{data.frame} with columns \code{"idx_tbls"} with database tables,
and \code{"idx_fields"} containing the field indexed for each table.
}
\description{
Get SQLite table index information
}
\details{
This function returns a summary of table index data, derived
from the SQL used to create the index. As such, it recognizes
most but not all possible examples of valid syntax. It currently
does not support compound column indexes, but this limitation
may change in future, or upon request.

Note that this function strictly queries the \code{sqlite_master} table
and parses the SQL for \code{type = 'index'}, and therefore will not
return information about database views. For that information,
try either \code{sqlite_view()} or \code{sqlite_desc()} both of which
indicate whether each field is associated with a column index.
}
\examples{
db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:");
DBI::dbCreateTable(db, "forecasts", data.frame(id=1:3, temperature=20:22));
DBI::dbGetQuery(db, "CREATE INDEX id_forecast_idx on forecasts(id)");
sqlite_indices(db, "forecasts")

}
\seealso{
Other jam database functions: \code{\link{ephys_short_name}},
  \code{\link{extract_event_data}},
  \code{\link{get_animal_event_data}},
  \code{\link{get_animal_event_derived}},
  \code{\link{get_animal_raw_data}},
  \code{\link{save_animal_event_data}},
  \code{\link{save_animal_event_derived}},
  \code{\link{save_animal_raw_data}},
  \code{\link{sqlite_desc}}, \code{\link{sqlite_view}}
}
\concept{jam database functions}
